+++ src/pages/safe/balances/nfts.tsx
@@ -9,7 +9,7 @@ import AssetsIcon from '@/public/images/sidebar/assets.svg'
 import NavTabs from '@/components/common/NavTabs'
 import { balancesNavItems } from '@/components/sidebar/SidebarNavigation/config'
 import ErrorMessage from '@/components/tx/ErrorMessage'
-import LoadMoreButton from '@/components/common/LoadMoreButton'
+import InfiniteScroll from '@/components/common/InfiniteScroll'
 
 const NftPage = ({
   pageUrl,
@@ -18,7 +18,7 @@ const NftPage = ({
   pageUrl: string
   onNextPage?: (pageUrl?: string) => void
 }): ReactElement => {
-  const [collectibles, error, loading] = useCollectibles(pageUrl)
+  const [collectibles, error] = useCollectibles(pageUrl)
 
   if (collectibles) {
     return (
@@ -33,7 +33,7 @@ const NftPage = ({
 
         {onNextPage && collectibles.next && (
           <Box py={4} textAlign="center">
-            <LoadMoreButton onLoadMore={() => onNextPage(collectibles.next)} loading={loading} />
+            <InfiniteScroll onLoadMore={() => onNextPage(collectibles.next)} />
           </Box>
         )}
       </>
